requirements.txt
ai-api-client-sdk==2.6.1
ai-core-sdk==2.6.2
fastapi==0.115.12
firebase-admin==6.8.0
generative-ai-hub-sdk==4.12.4
google-ai-generativelanguage==0.6.6
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.33.0
gradio_client==1.10.2
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.3.65
langchain-google-firestore==0.5.0
langchain-google-genai==2.1.4
langchain-openai==0.3.18
langchain-text-splitters==0.2.4
langsmith==0.3.45
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.82.0
regex==2024.11.6
tiktoken==0.7.0
tokenizers==0.20.3
websocket-client==1.8.0
websockets==15.0.1
python-dotenv==1.0.1
python==3.10.0



